+++ spec/transliteration_spec.rb
@@ -824,6 +824,8 @@ describe Ruby2JS do
         must_equal 'A = function() {var B = 1; return {B: B}}()'
       to_js( 'module A; def b; return 1; end; end' ).
         must_equal 'var A = {get b() {return 1}}'
+      to_js( 'module A; def b(); return 1; end; end' ).
+        must_equal 'var A = {b: function() {return 1}}'
       to_js( 'module A; class B; def initialize; @c=1; end; end; end' ).
         must_equal 'A = function() {function B() {this._c = 1}; return {B: B}}()'
     end