+++ lib/ruby2js.rb
@@ -10,6 +10,7 @@ end
 
 require 'ruby2js/converter'
 require 'ruby2js/filter'
+require 'ruby2js/namespace'
 
 module Ruby2JS
   class SyntaxError < RuntimeError
@@ -66,7 +67,7 @@ module Ruby2JS
       include Ruby2JS::Filter
       BINARY_OPERATORS = Converter::OPERATORS[2..-1].flatten
 
-      attr_accessor :prepend_list, :disable_autoimports
+      attr_accessor :prepend_list, :disable_autoimports, :namespace
 
       def initialize(comments)
         @comments = comments
@@ -207,6 +208,8 @@ module Ruby2JS
       comments = ast ? Parser::Source::Comment.associate(ast, comments) : {}
     end
 
+    namespace = Namespace.new
+
     filters = (options[:filters] || Filter::DEFAULTS)
 
     unless filters.empty?
@@ -221,6 +224,7 @@ module Ruby2JS
       filter = filter.new(comments)
 
       filter.options = options
+      filter.namespace = namespace
       ast = filter.process(ast)
 
       unless filter.prepend_list.empty?
@@ -240,6 +244,9 @@ module Ruby2JS
     ruby2js.or = options[:or] || :logical
     ruby2js.module_type = options[:module] || :esm
     ruby2js.underscored_private = (options[:eslevel] < 2020) || options[:underscored_private]
+
+    ruby2js.namespace = namespace
+
     if ruby2js.binding and not ruby2js.ivars
       ruby2js.ivars = ruby2js.binding.eval \
         'Hash[instance_variables.map {|var| [var, instance_variable_get(var)]}]'