+++ lib/ruby2js/converter.rb
@@ -1,5 +1,4 @@
 require 'ruby2js/serializer'
-require 'ruby2js/namespace'
 
 module Ruby2JS
   class Error < NotImplementedError
@@ -35,7 +34,7 @@ module Ruby2JS
 
     VASGN = [:cvasgn, :ivasgn, :gvasgn, :lvasgn]
 
-    attr_accessor :binding, :ivars
+    attr_accessor :binding, :ivars, :namespace
 
     def initialize( ast, comments, vars = {} )
       super()
@@ -51,8 +50,6 @@ module Ruby2JS
       @@handlers.each do |name|
         @handlers[name] = method("on_#{name}")
       end
-
-      @namespace = Namespace.new
 
       @state = nil
       @block_this = nil