+++ lib/ruby2js/converter/hide.rb
@@ -1,0 +1,12 @@
+module Ruby2JS
+  class Converter
+
+    # (hide, ...)
+
+    handle :hide do |*nodes|
+      capture {parse_all *nodes}
+
+      @lines.pop if @state == :statement and @lines.last == []
+    end
+  end
+end