+++ spec/require_spec.rb
@@ -1,6 +1,7 @@
 gem 'minitest'
 require 'minitest/autorun'
 require 'ruby2js/filter/require'
+require 'ruby2js/filter/esm'
 
 describe Ruby2JS::Filter::Require do
   
@@ -11,6 +12,26 @@ describe Ruby2JS::Filter::Require do
   
   def to_js(string)
     _(to_js_bare(string).to_s)
+  end
+  
+  def to_js_esm(string)
+    _(Ruby2JS.convert(string, 
+      filters: [Ruby2JS::Filter::Require, Ruby2JS::Filter::ESM],
+      file: __FILE__).to_s)
+  end
+  
+  def to_js_esm_auto(string)
+    _(Ruby2JS.convert(string, 
+      autoexports: true,
+      filters: [Ruby2JS::Filter::Require, Ruby2JS::Filter::ESM],
+      file: __FILE__).to_s)
+  end
+  
+  def to_js_esm_autodefault(string)
+    _(Ruby2JS.convert(string, 
+      autoexports: :default,
+      filters: [Ruby2JS::Filter::Require, Ruby2JS::Filter::ESM],
+      file: __FILE__).to_s)
   end
   
   describe :statement do
@@ -37,6 +58,23 @@ describe Ruby2JS::Filter::Require do
       _(timestamps[test1]).must_equal File.mtime(test1)
       _(timestamps[test2]).must_equal File.mtime(test2)
       _(timestamps[test3]).must_equal File.mtime(test3)
+    end
+  end
+
+  describe :esmimport do
+    it "should handle explicit exports" do
+      to_js_esm( 'require "require/test4.rb"' ).
+        must_equal 'import { Foo } from "require/test4.rb"; '
+    end
+
+    it "should handle auto exports" do
+      to_js_esm_auto( 'require "require/test5.rb"' ).
+        must_equal 'import { Foo } from "require/test5.rb"; '
+    end
+
+    it "should handle auto exports default" do
+      to_js_esm_autodefault( 'require "require/test5.rb"' ).
+        must_equal 'import Foo from "require/test5.rb"; '
     end
   end