+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Support static method calls with blocks in es2015+
 * Auto-bind instance methods referenced as properties within a class
 * handle begin, if, and case as expressions
+* handle modules with exactly one method
 
 # 3.6.1 / 2020-12-31