+++ CHANGELOG.md
@@ -2,8 +2,10 @@
 
 * Support static method calls with blocks in es2015+
 * Auto-bind instance methods referenced as properties within a class
-* handle begin, if, and case as expressions
-* handle modules with exactly one method
+* Handle begin, if, and case as expressions
+* Handle modules with exactly one method
+* Live demo based on Opal ([see here](https://ruby2js.com/demo))
+* Handle empty edge cases like `` `#{}` `` and `()`
 
 # 3.6.1 / 2020-12-31