+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Anonymous classes via Class.new
 * Autoexport :default option
 * Support both default and named imports on same import statement
+* requires for modules containing exports statements generate import statements
 
 # 3.6.1 / 2020-12-31